move cleanning state to separate mod
