+++ src/cli/cleanup/app/mod.rs
@@ -1,13 +1,11 @@
+mod cleanning;
 mod ui;
 
-use std::{
-    collections::HashMap,
-    hash::Hash,
-    sync::mpsc::{Receiver, RecvTimeoutError},
-};
+use std::sync::mpsc::Receiver;
 
 use ratatui::crossterm::event::{self, Event, KeyCode, KeyEventKind};
 
+use self::cleanning::CleanningState;
 use crate::{
     buildkit::{BuildKitD, container_info::SCellContainerInfo, image_info::SCellImageInfo},
     cli::{MIN_FPS, terminal::Terminal},
@@ -18,8 +16,8 @@ pub enum App {
         rx: Receiver<color_eyre::Result<(Vec<SCellContainerInfo>, Vec<SCellImageInfo>)>>,
         buildkit: BuildKitD,
     },
-    CleaningContainers(CleaningState<SCellContainerInfo>),
-    CleaningImages(CleaningState<SCellImageInfo>),
+    CleanningContainers(CleanningState<SCellContainerInfo>),
+    CleanningImages(CleanningState<SCellImageInfo>),
     Exit,
 }
 
@@ -42,17 +40,17 @@ impl App {
             {
                 let (containers_for_removal, images_for_removal_res) = result?;
                 images_for_removal = images_for_removal_res;
-                app = Self::cleaning_containers(containers_for_removal, buildkit.clone());
+                app = CleanningState::cleaning_containers(containers_for_removal, buildkit.clone());
             }
 
-            if let App::CleaningContainers(ref mut state) = app
+            if let App::CleanningContainers(ref mut state) = app
                 && state.try_update()
             {
                 let images_for_removal = std::mem::take(&mut images_for_removal);
-                app = Self::cleaning_images(images_for_removal, buildkit.clone());
+                app = CleanningState::cleaning_images(images_for_removal, buildkit.clone());
             }
 
-            if let App::CleaningImages(ref mut state) = app
+            if let App::CleanningImages(ref mut state) = app
                 && state.try_update()
             {
                 app = App::Exit;
@@ -104,50 +102,6 @@ impl App {
         App::Loading { rx, buildkit }
     }
 
-    fn cleaning_containers(
-        for_removal: Vec<SCellContainerInfo>,
-        buildkit: BuildKitD,
-    ) -> Self {
-        let (tx, rx) = std::sync::mpsc::channel();
-
-        // Spawn async task to cleanup orphan containers and their corresponding images
-        tokio::spawn({
-            let containers = for_removal.clone();
-            async move {
-                for c in containers {
-                    let res = buildkit.cleanup_container(&c).await;
-                    tokio::time::sleep(std::time::Duration::from_millis(300)).await;
-                    drop(tx.send((c, res)));
-                }
-                tokio::time::sleep(std::time::Duration::from_millis(500)).await;
-            }
-        });
-
-        App::CleaningContainers(CleaningState::new(for_removal, rx))
-    }
-
-    fn cleaning_images(
-        for_removal: Vec<SCellImageInfo>,
-        buildkit: BuildKitD,
-    ) -> Self {
-        let (tx, rx) = std::sync::mpsc::channel();
-
-        // Spawn async task to cleanup orphan images
-        tokio::spawn({
-            let images = for_removal.clone();
-            async move {
-                for c in images {
-                    let res = buildkit.cleanup_image(&c).await;
-                    tokio::time::sleep(std::time::Duration::from_millis(300)).await;
-                    drop(tx.send((c, res)));
-                }
-                tokio::time::sleep(std::time::Duration::from_millis(500)).await;
-            }
-        });
-
-        App::CleaningImages(CleaningState::new(for_removal, rx))
-    }
-
     fn handle_key_event(mut self) -> color_eyre::Result<Self> {
         if event::poll(MIN_FPS)?
             && let Event::Key(key) = event::read()?
@@ -159,34 +113,5 @@ impl App {
         }
 
         Ok(self)
-    }
-}
-
-pub struct CleaningState<Item> {
-    removing_results: HashMap<Item, Option<color_eyre::Result<()>>>,
-    rx: Receiver<(Item, color_eyre::Result<()>)>,
-}
-
-impl<Item: Clone + PartialEq + Eq + Hash> CleaningState<Item> {
-    pub fn new(
-        for_removal: Vec<Item>,
-        rx: Receiver<(Item, color_eyre::Result<()>)>,
-    ) -> Self {
-        Self {
-            removing_results: for_removal.into_iter().map(|c| (c, None)).collect(),
-            rx,
-        }
-    }
-
-    /// Returns boolean flag, if the udelrying channel was closed or not
-    fn try_update(&mut self) -> bool {
-        match self.rx.recv_timeout(MIN_FPS) {
-            Ok(update) => {
-                self.removing_results.insert(update.0, Some(update.1));
-                false
-            },
-            Err(RecvTimeoutError::Timeout) => false,
-            Err(RecvTimeoutError::Disconnected) => true,
-        }
     }
 }
